Create rpc-with-autoplayupdate.js
This code is an updated version of rpc.js .
In this code one new button is created 'Auto Play' which will play the game with its own by clicking on it. 
The game will automatically turn off by clicking the button again.
Functions like setInterval and clearInterval are used to update the score in every 2 seconds. 
Make sure to create a button in the HTML file and give it the class mentioned in this code.