Create rpc-with-eventListener.js
This is an updated version of rock paper scissors game where addeventlistener property is used. when a user clicks on a button as r,p or s it automatically calls the playgame function with rock paper scissor as arguments.
note : please add a class to rock paper scissor buttons as the name used in addeventlistener property in this js file.